fixing redstone networking, add spread and in aerosolize modes for launcher
Fixing some of the redstone networking issues with a bit more logic. Added a spreadmode to the launcher for quivers, and now allow you to use an aerosolize  funtion the same as the mages spell, but with an option for a expanded radius.

Also gave the guildmaster a master wrench, one that can link multiple redstone structures together.